+++ packages/Charts/src/viewers/tree/tree-viewer.ts
@@ -602,6 +602,7 @@ export class TreeViewer extends EChartViewer {
 
   addSubs(): void {
     if (!this.dataFrame) return;
+    this.subs.push(DG.debounce(this.dataFrame.onDataChanged, 50).subscribe((_) => this.render()));
     this.subs.push(this.dataFrame.onColumnsRemoved.subscribe((data) => {
       const columnNamesToRemove = data.columns.map((column: DG.Column) => column.name);
       this.hierarchyColumnNames = this.hierarchyColumnNames.filter((columnName) => !columnNamesToRemove.includes(columnName));
@@ -920,13 +921,14 @@ export class TreeViewer extends EChartViewer {
     if (this.chart) {
       this.chart.clear();
       this.chart.dispose();
-      this.detach();
       this.chart = null;
     }
 
     this.chart = echarts.init(this.root);
-    this.initChartEventListeners();
-    this.addSubs();
+    this.resubscribe(() => {
+      this.initChartEventListeners();
+      this.addSubs();
+    });
 
     this.option.series[0].label.formatter = (params: any) => this.formatLabel(params);
     this.chart.setOption(this.option, false, true);